refactor: extract gather/break embeds into embeds/ folder
Move the gathering and break embed builders out of cmd_gather, cmd_break
and gather_service into dedicated embeds/gather_embed.rs and
embeds/break_embed.rs, matching the project's existing convention where
every domain owns a single embed module. Also lift BreakState and the
break-timer loop into a new service/break_service.rs so the break flow
mirrors gather_service, leaving the command modules as thin shells that
only parse args and dispatch.
