+++ src/embeds/mod.rs
@@ -1,4 +1,6 @@
 pub mod bot_embeds;
+pub mod break_embed;
+pub mod gather_embed;
 pub mod notify_embeds;
 pub mod player_embed;
 pub mod queue_embed;
