+++ src/service/mod.rs
@@ -1,3 +1,4 @@
+pub mod break_service;
 pub mod cache_service;
 pub mod channel_service;
 pub mod embed_service;
